プログラムとテストコードの整理
