+++ glisp/src/lib.rs
@@ -22,6 +22,7 @@ mod tests {
 
     use std::fs::File;
     use std::io::Write;
+    use std::process::Command;
     use std::time::{SystemTime, UNIX_EPOCH};
 
     fn do_lisp_env(program: &str, env: &Environment) -> String {
@@ -98,9 +99,24 @@ mod tests {
         assert_eq!(do_lisp_env("(screen-width)", &env), "720");
         assert_eq!(do_lisp_env("(screen-height)", &env), "560");
         assert_eq!(do_lisp_env("(draw-eval (iota 10))", &env), "nil");
-        assert_eq!(do_lisp_env("(gtk-major-version)", &env), "3");
-        assert_eq!(do_lisp_env("(gtk-minor-version)", &env), "22");
-        assert_eq!(do_lisp_env("(gtk-micro-version)", &env), "30");
+
+        let output = Command::new("pkg-config")
+            .arg("--modversion")
+            .arg("gtk+-3.0")
+            .output()
+            .expect("gtk-xx-version falut");
+        let version = String::from_utf8(output.stdout).unwrap();
+        let version = version.split(".");
+        let mut iter = version.into_iter();
+        if let Some(v) = iter.next() {
+            assert_eq!(do_lisp_env("(gtk-major-version)", &env), v);
+        }
+        if let Some(v) = iter.next() {
+            assert_eq!(do_lisp_env("(gtk-minor-version)", &env), v);
+        }
+        if let Some(v) = iter.next() {
+            assert_eq!(do_lisp_env("(gtk-micro-version)", &env), v.trim_end());
+        }
     }
     #[test]
     fn test_02_error_check() {
