+++ glisp/src/ui.rs
@@ -15,8 +15,6 @@ use lisp::Environment;
 
 use gtk::prelude::*;
 use std::env;
-use std::fs;
-use std::path::PathBuf;
 use std::rc::Rc;
 
 use crate::draw::draw_clear;
@@ -24,6 +22,7 @@ use crate::draw::get_default_surface;
 use crate::draw::save_png_file;
 use crate::draw::DrawTable;
 use crate::draw::Graffiti;
+use crate::helper::load_demo_program;
 use crate::helper::search_word_highlight;
 use crate::helper::History;
 use crate::helper::SourceView;
@@ -40,7 +39,6 @@ const DRAW_CLEAR_KEYCODE: u32 = 108;
 
 #[cfg(feature = "animation")]
 const MOTION_DELAY: i32 = 70;
-
 const STOP_ERROR_CODE: &str = "E9000";
 
 //--------------------------------------------------------
@@ -136,46 +134,6 @@ textview {
         &provider,
         gtk::STYLE_PROVIDER_PRIORITY_APPLICATION,
     );
-}
-//--------------------------------------------------------
-// Load LISP programe(https://github.com/hidekuno/picture-language)
-//--------------------------------------------------------
-fn load_demo_program(dir: &str) -> std::io::Result<String> {
-    fn get_program_name(vec: Vec<&str>) -> std::io::Result<Option<String>> {
-        let mut program: Vec<String> = Vec::new();
-        let mut path = PathBuf::new();
-
-        path.push(match env::var("HOME") {
-            Ok(v) => v,
-            Err(_) => "/root".into(),
-        });
-        for dir in vec {
-            path.push(dir);
-        }
-        if false == path.as_path().exists() {
-            return Ok(None);
-        }
-        for entry in fs::read_dir(path)? {
-            let dir = entry?;
-            let path = dir.path();
-            let f = path.to_str().unwrap();
-            if f.ends_with(".scm") {
-                program.push(format!("(load-file \"{}\")", f));
-            }
-        }
-        program.sort();
-        Ok(Some(program.join("\n")))
-    }
-    for v in vec![vec!["picture-language", dir], vec![dir]] {
-        match get_program_name(v) {
-            Ok(s) => match s {
-                Some(s) => return Ok(s),
-                None => continue,
-            },
-            Err(e) => return Err(e),
-        }
-    }
-    Ok("".into())
 }
 //--------------------------------------------------------
 // SICP Program
